% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptors.R
\name{getDockerWorkerNumbers,ManagedCloudProvider-method}
\alias{getDockerWorkerNumbers,ManagedCloudProvider-method}
\title{get the worker number in the cluster}
\usage{
\S4method{getDockerWorkerNumbers}{ManagedCloudProvider}(provider, cluster, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A list with \code{initializing} and \code{running} integer elements
}
\description{
The developer should define \code{runDockerWorkerContainers}, \code{getDockerWorkerStatus}
and \code{killDockerWorkerContainers} instead.
}
