% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-generics.R, R/dummyManagedProvider.R
\name{killDockerWorkerContainers}
\alias{killDockerWorkerContainers}
\alias{killDockerWorkerContainers,DummyManagedProvider-method}
\title{Kill the worker container}
\usage{
killDockerWorkerContainers(provider, cluster, workerHandles, verbose)

\S4method{killDockerWorkerContainers}{DummyManagedProvider}(provider, cluster, workerHandles, verbose)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerHandles}{Character(n). A character vector of \strong{unique} instance handles.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A logical vector indicating whether the killing operation is successful for each instance
}
\description{
Kill the worker container. The worker handles are unique.
If multiple workers share the same instance, all workers in the
same container should be killed. There is no default method for this generic.
}
\section{Functions}{
\itemize{
\item \code{killDockerWorkerContainers,DummyManagedProvider-method}: The method for the dummy managed provider
}}

