% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptors.R
\name{setDockerWorkerNumber,ManagedCloudProvider-method}
\alias{setDockerWorkerNumber,ManagedCloudProvider-method}
\title{Set the worker number in the cluster}
\usage{
\S4method{setDockerWorkerNumber}{ManagedCloudProvider}(
  provider,
  cluster,
  container,
  hardware,
  workerNumber,
  verbose
)
}
\arguments{
\item{provider}{S4 \code{CloudProvider} object. The service provider.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{container}{S4 \code{DockerContainer} Object.}

\item{hardware}{S4 \code{DockerHardware} Object.}

\item{workerNumber}{Integer(1), the worker number to be set}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
No return value
}
\description{
The developer should define \code{runDockerWorkerContainers}, \code{getDockerWorkerStatus}
and \code{killDockerWorkerContainers} instead.
}
